cursed momentum

+++ src/main/java/samuschair/orbital2/window/SpinnyCircleWindow.java
@@ -28,8 +28,8 @@ public class SpinnyCircleWindow extends Window {
 		NkRect space = NkRect.malloc(stack);
 		nk_widget(space, ctx);
 
-		float x = (float) (space.x() + Math.cos(motion_X * 0.1) * 50) + width / 2 - 20;
-		float y = (float) (space.y()  + Math.sin(motion_Y * 0.1) * 50) + height / 2 - 20;
+		float x = (float) (space.x() + Math.cos(motion_X * 0.1) * 50) + ((float) width / 2) - 20;
+		float y = (float) (space.y()  + Math.sin(motion_Y * 0.1) * 50) + ((float) height / 2) - 20;
 		nk_fill_circle(canvas, nk_rect(x, y, 40, 40, space), nk_rgb(50, 50, 220, NkColor.malloc(stack)));
 		motion_X++;
 		motion_Y++;
